got customers moving a little bit
also removed the rotation target for the player, and just had it face the mouse. the slerp was causing issues and it moved just about as fast as just moving directly to where the mouse is, so we're gonna do this for now until we can figure out how to slerp properly.