move some classes into their own modules
trying to reduce all the code in a single file thing, not a good luck, huh?

also finally got this usable in a live preview in visual studio code, which is great. i've just been using tinyfilemanager to directly manage the files on the web, and it was a huge pain the ass.

turns out i can bypass CORS using htaccess.