fix up kinematic objects
Kinematic cube is now just a RigidBodyCube with a mass of zero. Which, now that I think about it, is what it already was. But it works.

Added a setPosition function to PhysicsObject that seems to work well for now. Took a minute to figure it out though.

Player is a RigidBodyCube that gets forced applied to it, though it definitely needs work done to it to make it smoother.