fixes & improvements
fixed rep not going down when customers leave due to time by having customers keep track of their moods over time, and then adding the mood to the rep when the customer leaves.

also moved the customer overtime leave code into the customer class and out of the shop class.

started work to improve the interface