improved physics and more car work
RigidBodyCubeComponent is now RigidBodyComponent. It auto adds a geometry component instead of handling geometry itself. I stated adding the groundwork for it automatically choose the physics geometry type based on the mesh type, so that spheres, cylinders, and hopefully also models can be supported.

Added fullbright mode which replaces the clear color with white.